Refactor common functions into top-level utility module. [ci skip]
